Backend/python/langchain/requirements.txt
# Modern LangChain Agent System Dependencies

# Core framework
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
starlette

# LangChain components
langchain
langchain-core
langchain-openai
langchain-text-splitters
langsmith

# Azure services
azure-ai-projects
azure-ai-inference
azure-identity
azure-core
azure-search-documents
azure-cosmos
azure-common

# OpenAI
openai
tiktoken

# Async and utilities
aiofiles
aioredis  # Optional for Redis session storage
httpx
httpcore
anyio
aiohttp

# Configuration and validation
PyYAML
jsonpatch
jsonpointer
typing-extensions>=4.8.0  # From shared

# Session storage (optional)
redis>=4.5.0  # For Redis session storage
sqlalchemy>=2.0.0  # For database session storage

# Shared dependencies
# Note: Shared module is imported via sys.path modification in the code
# No need for separate installation

# Development (optional)
pytest
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
mypy>=1.6.0
pre-commit>=3.0.0
pytest-asyncio
black
isort
mypy

# Optional session storage backends
redis  # For Redis session storage
sqlalchemy  # For database session storage

# Additional dependencies for text processing
regex
orjson
tenacity